+++ packages/ui-svelte/src/index.ts
@@ -130,6 +130,7 @@ export { default as BreadcrumbSeparator } from './lib/components/BreadcrumbSepar
 
 // Calendar
 export { default as Calendar } from './lib/components/Calendar.svelte'
+export { default as DatePicker } from './lib/components/DatePicker.svelte'
 
 // Carousel
 export { default as Carousel } from './lib/components/Carousel.svelte'